Don't escape the scorer names
bastet writes the scores file using double-quotes around the name for some reason (probably to avoid some kind of injection?), but reads the file without removing the double-quotes. The scorer name then appears as `"tester"` instead of `tester` in the scoreboard, quitting the game makes bastet write `""tester""` as name. So every time the game scorer file is written, double-quotes get added to the scorer name.

Since bastet reads the scorer name from the scorer file without removing the double-quotes or (un-)escaping anything, the easiest way to work around this is to make bastet not write double-quotes around the scorer name.